refactor: modularize server and fix build
